efactor: 프로젝트 구조를 파일별로 모듈화하여 개선
