+++ src/main/kotlin/study/dsl/Builders.kt
@@ -1,0 +1,57 @@
+package study.dsl
+
+class PersonBuilder {
+    private lateinit var name: String
+    private var company: String? = null
+    private var skills: Skills? = null
+    private var languages: Languages? = null
+
+    fun name(value: String) {
+        name = value
+    }
+
+    fun company(value: String) {
+        company = value
+    }
+
+    fun skills(block: SkillsBuilder.() -> Unit) {
+        skills = SkillsBuilder().apply(block).build()
+    }
+
+    fun languages(block: LanguageBuilder.() -> Unit) {
+        languages = LanguageBuilder().apply(block).build()
+    }
+
+    fun build(): Person {
+        return Person(name, company, skills, languages)
+    }
+}
+
+class SkillsBuilder {
+    private val soft = mutableListOf<String>()
+    private val hard = mutableListOf<String>()
+
+    fun soft(value: String) {
+        soft.add(value)
+    }
+
+    fun hard(value: String) {
+        hard.add(value)
+    }
+
+    fun build(): Skills {
+        return Skills(soft, hard)
+    }
+}
+
+class LanguageBuilder {
+    private val languages = mutableMapOf<String, Int>()
+
+    infix fun String.level(level: Int) {
+        languages[this] = level
+    }
+
+    fun build(): Languages {
+        return Languages(languages)
+    }
+}
