+++ src/main/kotlin/study/dsl/Dsl.kt
@@ -1,0 +1,5 @@
+package study.dsl
+
+fun introduce(block: PersonBuilder.() -> Unit): Person {
+    return PersonBuilder().apply(block).build()
+}
